requirements.txt
fastapi
uvicorn
sentence-transformers
torch
numpy
faiss-cpu
scikit-learn
python-dotenv
